Move base registry into separate file
