+++ sdk/python/feast/infra/materialization/batch_materialization_engine.py
@@ -11,7 +11,7 @@ from feast.entity import Entity
 from feast.feature_view import FeatureView
 from feast.infra.offline_stores.offline_store import OfflineStore
 from feast.infra.online_stores.online_store import OnlineStore
-from feast.infra.registry.registry import BaseRegistry
+from feast.infra.registry.base_registry import BaseRegistry
 from feast.repo_config import RepoConfig
 from feast.stream_feature_view import StreamFeatureView
