+++ sdk/python/feast/infra/registry/registry.py
@@ -11,12 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import abc
-import json
 import logging
 import uuid
-from abc import abstractmethod
-from collections import defaultdict
 from datetime import datetime, timedelta
 from enum import Enum
 from pathlib import Path
@@ -25,7 +21,6 @@ from typing import Any, Dict, List, Optional
 from urllib.parse import urlparse
 
 from google.protobuf.internal.containers import RepeatedCompositeFieldContainer
-from google.protobuf.json_format import MessageToJson
 from proto import Message
 
 from feast import usage
@@ -47,6 +42,7 @@ from feast.feature_service import FeatureService
 from feast.feature_view import FeatureView
 from feast.importer import import_class
 from feast.infra.infra_object import Infra
+from feast.infra.registry.base_registry import BaseRegistry
 from feast.infra.registry.registry_store import NoopRegistryStore
 from feast.on_demand_feature_view import OnDemandFeatureView
 from feast.project_metadata import ProjectMetadata
@@ -148,614 +144,6 @@ def get_registry_store_class_from_scheme(registry_path: str):
         return get_registry_store_class_from_type(registry_store_type)
 
 
-class BaseRegistry(abc.ABC):
-    # Entity operations
-    @abstractmethod
-    def apply_entity(self, entity: Entity, project: str, commit: bool = True):
-        """
-        Registers a single entity with Feast
-
-        Args:
-            entity: Entity that will be registered
-            project: Feast project that this entity belongs to
-            commit: Whether the change should be persisted immediately
-        """
-
-    @abstractmethod
-    def delete_entity(self, name: str, project: str, commit: bool = True):
-        """
-        Deletes an entity or raises an exception if not found.
-
-        Args:
-            name: Name of entity
-            project: Feast project that this entity belongs to
-            commit: Whether the change should be persisted immediately
-        """
-
-    @abstractmethod
-    def get_entity(self, name: str, project: str, allow_cache: bool = False) -> Entity:
-        """
-        Retrieves an entity.
-
-        Args:
-            name: Name of entity
-            project: Feast project that this entity belongs to
-            allow_cache: Whether to allow returning this entity from a cached registry
-
-        Returns:
-            Returns either the specified entity, or raises an exception if
-            none is found
-        """
-
-    @abstractmethod
-    def list_entities(self, project: str, allow_cache: bool = False) -> List[Entity]:
-        """
-        Retrieve a list of entities from the registry
-
-        Args:
-            allow_cache: Whether to allow returning entities from a cached registry
-            project: Filter entities based on project name
-
-        Returns:
-            List of entities
-        """
-
-    # Data source operations
-    @abstractmethod
-    def apply_data_source(
-        self, data_source: DataSource, project: str, commit: bool = True
-    ):
-        """
-        Registers a single data source with Feast
-
-        Args:
-            data_source: A data source that will be registered
-            project: Feast project that this data source belongs to
-            commit: Whether to immediately commit to the registry
-        """
-
-    @abstractmethod
-    def delete_data_source(self, name: str, project: str, commit: bool = True):
-        """
-        Deletes a data source or raises an exception if not found.
-
-        Args:
-            name: Name of data source
-            project: Feast project that this data source belongs to
-            commit: Whether the change should be persisted immediately
-        """
-
-    @abstractmethod
-    def get_data_source(
-        self, name: str, project: str, allow_cache: bool = False
-    ) -> DataSource:
-        """
-        Retrieves a data source.
-
-        Args:
-            name: Name of data source
-            project: Feast project that this data source belongs to
-            allow_cache: Whether to allow returning this data source from a cached registry
-
-        Returns:
-            Returns either the specified data source, or raises an exception if none is found
-        """
-
-    @abstractmethod
-    def list_data_sources(
-        self, project: str, allow_cache: bool = False
-    ) -> List[DataSource]:
-        """
-        Retrieve a list of data sources from the registry
-
-        Args:
-            project: Filter data source based on project name
-            allow_cache: Whether to allow returning data sources from a cached registry
-
-        Returns:
-            List of data sources
-        """
-
-    # Feature service operations
-    @abstractmethod
-    def apply_feature_service(
-        self, feature_service: FeatureService, project: str, commit: bool = True
-    ):
-        """
-        Registers a single feature service with Feast
-
-        Args:
-            feature_service: A feature service that will be registered
-            project: Feast project that this entity belongs to
-        """
-
-    @abstractmethod
-    def delete_feature_service(self, name: str, project: str, commit: bool = True):
-        """
-        Deletes a feature service or raises an exception if not found.
-
-        Args:
-            name: Name of feature service
-            project: Feast project that this feature service belongs to
-            commit: Whether the change should be persisted immediately
-        """
-
-    @abstractmethod
-    def get_feature_service(
-        self, name: str, project: str, allow_cache: bool = False
-    ) -> FeatureService:
-        """
-        Retrieves a feature service.
-
-        Args:
-            name: Name of feature service
-            project: Feast project that this feature service belongs to
-            allow_cache: Whether to allow returning this feature service from a cached registry
-
-        Returns:
-            Returns either the specified feature service, or raises an exception if
-            none is found
-        """
-
-    @abstractmethod
-    def list_feature_services(
-        self, project: str, allow_cache: bool = False
-    ) -> List[FeatureService]:
-        """
-        Retrieve a list of feature services from the registry
-
-        Args:
-            allow_cache: Whether to allow returning entities from a cached registry
-            project: Filter entities based on project name
-
-        Returns:
-            List of feature services
-        """
-
-    # Feature view operations
-    @abstractmethod
-    def apply_feature_view(
-        self, feature_view: BaseFeatureView, project: str, commit: bool = True
-    ):
-        """
-        Registers a single feature view with Feast
-
-        Args:
-            feature_view: Feature view that will be registered
-            project: Feast project that this feature view belongs to
-            commit: Whether the change should be persisted immediately
-        """
-
-    @abstractmethod
-    def delete_feature_view(self, name: str, project: str, commit: bool = True):
-        """
-        Deletes a feature view or raises an exception if not found.
-
-        Args:
-            name: Name of feature view
-            project: Feast project that this feature view belongs to
-            commit: Whether the change should be persisted immediately
-        """
-
-    # stream feature view operations
-    @abstractmethod
-    def get_stream_feature_view(
-        self, name: str, project: str, allow_cache: bool = False
-    ):
-        """
-        Retrieves a stream feature view.
-
-        Args:
-            name: Name of stream feature view
-            project: Feast project that this feature view belongs to
-            allow_cache: Allow returning feature view from the cached registry
-
-        Returns:
-            Returns either the specified feature view, or raises an exception if
-            none is found
-        """
-
-    @abstractmethod
-    def list_stream_feature_views(
-        self, project: str, allow_cache: bool = False
-    ) -> List[StreamFeatureView]:
-        """
-        Retrieve a list of stream feature views from the registry
-
-        Args:
-            project: Filter stream feature views based on project name
-            allow_cache: Whether to allow returning stream feature views from a cached registry
-
-        Returns:
-            List of stream feature views
-        """
-
-    # on demand feature view operations
-    @abstractmethod
-    def get_on_demand_feature_view(
-        self, name: str, project: str, allow_cache: bool = False
-    ) -> OnDemandFeatureView:
-        """
-        Retrieves an on demand feature view.
-
-        Args:
-            name: Name of on demand feature view
-            project: Feast project that this on demand feature view belongs to
-            allow_cache: Whether to allow returning this on demand feature view from a cached registry
-
-        Returns:
-            Returns either the specified on demand feature view, or raises an exception if
-            none is found
-        """
-
-    @abstractmethod
-    def list_on_demand_feature_views(
-        self, project: str, allow_cache: bool = False
-    ) -> List[OnDemandFeatureView]:
-        """
-        Retrieve a list of on demand feature views from the registry
-
-        Args:
-            project: Filter on demand feature views based on project name
-            allow_cache: Whether to allow returning on demand feature views from a cached registry
-
-        Returns:
-            List of on demand feature views
-        """
-
-    # regular feature view operations
-    @abstractmethod
-    def get_feature_view(
-        self, name: str, project: str, allow_cache: bool = False
-    ) -> FeatureView:
-        """
-        Retrieves a feature view.
-
-        Args:
-            name: Name of feature view
-            project: Feast project that this feature view belongs to
-            allow_cache: Allow returning feature view from the cached registry
-
-        Returns:
-            Returns either the specified feature view, or raises an exception if
-            none is found
-        """
-
-    @abstractmethod
-    def list_feature_views(
-        self, project: str, allow_cache: bool = False
-    ) -> List[FeatureView]:
-        """
-        Retrieve a list of feature views from the registry
-
-        Args:
-            allow_cache: Allow returning feature views from the cached registry
-            project: Filter feature views based on project name
-
-        Returns:
-            List of feature views
-        """
-
-    # request feature view operations
-    @abstractmethod
-    def get_request_feature_view(self, name: str, project: str) -> RequestFeatureView:
-        """
-        Retrieves a request feature view.
-
-        Args:
-            name: Name of request feature view
-            project: Feast project that this feature view belongs to
-            allow_cache: Allow returning feature view from the cached registry
-
-        Returns:
-            Returns either the specified feature view, or raises an exception if
-            none is found
-        """
-
-    @abstractmethod
-    def list_request_feature_views(
-        self, project: str, allow_cache: bool = False
-    ) -> List[RequestFeatureView]:
-        """
-        Retrieve a list of request feature views from the registry
-
-        Args:
-            allow_cache: Allow returning feature views from the cached registry
-            project: Filter feature views based on project name
-
-        Returns:
-            List of request feature views
-        """
-
-    @abstractmethod
-    def apply_materialization(
-        self,
-        feature_view: FeatureView,
-        project: str,
-        start_date: datetime,
-        end_date: datetime,
-        commit: bool = True,
-    ):
-        """
-        Updates materialization intervals tracked for a single feature view in Feast
-
-        Args:
-            feature_view: Feature view that will be updated with an additional materialization interval tracked
-            project: Feast project that this feature view belongs to
-            start_date (datetime): Start date of the materialization interval to track
-            end_date (datetime): End date of the materialization interval to track
-            commit: Whether the change should be persisted immediately
-        """
-
-    # Saved dataset operations
-    @abstractmethod
-    def apply_saved_dataset(
-        self,
-        saved_dataset: SavedDataset,
-        project: str,
-        commit: bool = True,
-    ):
-        """
-        Stores a saved dataset metadata with Feast
-
-        Args:
-            saved_dataset: SavedDataset that will be added / updated to registry
-            project: Feast project that this dataset belongs to
-            commit: Whether the change should be persisted immediately
-        """
-
-    @abstractmethod
-    def get_saved_dataset(
-        self, name: str, project: str, allow_cache: bool = False
-    ) -> SavedDataset:
-        """
-        Retrieves a saved dataset.
-
-        Args:
-            name: Name of dataset
-            project: Feast project that this dataset belongs to
-            allow_cache: Whether to allow returning this dataset from a cached registry
-
-        Returns:
-            Returns either the specified SavedDataset, or raises an exception if
-            none is found
-        """
-
-    def delete_saved_dataset(self, name: str, project: str, allow_cache: bool = False):
-        """
-        Delete a saved dataset.
-
-        Args:
-            name: Name of dataset
-            project: Feast project that this dataset belongs to
-            allow_cache: Whether to allow returning this dataset from a cached registry
-
-        Returns:
-            Returns either the specified SavedDataset, or raises an exception if
-            none is found
-        """
-
-    @abstractmethod
-    def list_saved_datasets(
-        self, project: str, allow_cache: bool = False
-    ) -> List[SavedDataset]:
-        """
-        Retrieves a list of all saved datasets in specified project
-
-        Args:
-            project: Feast project
-            allow_cache: Whether to allow returning this dataset from a cached registry
-
-        Returns:
-            Returns the list of SavedDatasets
-        """
-
-    # Validation reference operations
-    @abstractmethod
-    def apply_validation_reference(
-        self,
-        validation_reference: ValidationReference,
-        project: str,
-        commit: bool = True,
-    ):
-        """
-        Persist a validation reference
-
-        Args:
-            validation_reference: ValidationReference that will be added / updated to registry
-            project: Feast project that this dataset belongs to
-            commit: Whether the change should be persisted immediately
-        """
-
-    @abstractmethod
-    def delete_validation_reference(self, name: str, project: str, commit: bool = True):
-        """
-        Deletes a validation reference or raises an exception if not found.
-
-        Args:
-            name: Name of validation reference
-            project: Feast project that this object belongs to
-            commit: Whether the change should be persisted immediately
-        """
-
-    @abstractmethod
-    def get_validation_reference(
-        self, name: str, project: str, allow_cache: bool = False
-    ) -> ValidationReference:
-        """
-        Retrieves a validation reference.
-
-        Args:
-            name: Name of dataset
-            project: Feast project that this dataset belongs to
-            allow_cache: Whether to allow returning this dataset from a cached registry
-
-        Returns:
-            Returns either the specified ValidationReference, or raises an exception if
-            none is found
-        """
-
-    # TODO: Needs to be implemented.
-    def list_validation_references(
-        self, project: str, allow_cache: bool = False
-    ) -> List[ValidationReference]:
-
-        """
-        Retrieve a list of validation references from the registry
-
-        Args:
-            allow_cache: Allow returning feature views from the cached registry
-            project: Filter feature views based on project name
-
-        Returns:
-            List of request feature views
-        """
-
-    def list_project_metadata(
-        self, project: str, allow_cache: bool = False
-    ) -> List[ProjectMetadata]:
-        """
-        Retrieves project metadata
-
-        Args:
-            project: Filter metadata based on project name
-            allow_cache: Allow returning feature views from the cached registry
-
-        Returns:
-            List of project metadata
-        """
-
-    @abstractmethod
-    def update_infra(self, infra: Infra, project: str, commit: bool = True):
-        """
-        Updates the stored Infra object.
-
-        Args:
-            infra: The new Infra object to be stored.
-            project: Feast project that the Infra object refers to
-            commit: Whether the change should be persisted immediately
-        """
-
-    @abstractmethod
-    def get_infra(self, project: str, allow_cache: bool = False) -> Infra:
-        """
-        Retrieves the stored Infra object.
-
-        Args:
-            project: Feast project that the Infra object refers to
-            allow_cache: Whether to allow returning this entity from a cached registry
-
-        Returns:
-            The stored Infra object.
-        """
-
-    @abstractmethod
-    def apply_user_metadata(
-        self,
-        project: str,
-        feature_view: BaseFeatureView,
-        metadata_bytes: Optional[bytes],
-    ):
-        ...
-
-    @abstractmethod
-    def get_user_metadata(
-        self, project: str, feature_view: BaseFeatureView
-    ) -> Optional[bytes]:
-        ...
-
-    @abstractmethod
-    def proto(self) -> RegistryProto:
-        """
-        Retrieves a proto version of the registry.
-
-        Returns:
-            The registry proto object.
-        """
-
-    @abstractmethod
-    def commit(self):
-        """Commits the state of the registry cache to the remote registry store."""
-
-    @abstractmethod
-    def refresh(self, project: Optional[str]):
-        """Refreshes the state of the registry cache by fetching the registry state from the remote registry store."""
-
-    @staticmethod
-    def _message_to_sorted_dict(message: Message) -> Dict[str, Any]:
-        return json.loads(MessageToJson(message, sort_keys=True))
-
-    def to_dict(self, project: str) -> Dict[str, List[Any]]:
-        """Returns a dictionary representation of the registry contents for the specified project.
-
-        For each list in the dictionary, the elements are sorted by name, so this
-        method can be used to compare two registries.
-
-        Args:
-            project: Feast project to convert to a dict
-        """
-        registry_dict: Dict[str, Any] = defaultdict(list)
-        registry_dict["project"] = project
-        for project_metadata in sorted(self.list_project_metadata(project=project)):
-            registry_dict["projectMetadata"].append(
-                self._message_to_sorted_dict(project_metadata.to_proto())
-            )
-        for data_source in sorted(
-            self.list_data_sources(project=project), key=lambda ds: ds.name
-        ):
-            registry_dict["dataSources"].append(
-                self._message_to_sorted_dict(data_source.to_proto())
-            )
-        for entity in sorted(
-            self.list_entities(project=project), key=lambda entity: entity.name
-        ):
-            registry_dict["entities"].append(
-                self._message_to_sorted_dict(entity.to_proto())
-            )
-        for feature_view in sorted(
-            self.list_feature_views(project=project),
-            key=lambda feature_view: feature_view.name,
-        ):
-            registry_dict["featureViews"].append(
-                self._message_to_sorted_dict(feature_view.to_proto())
-            )
-        for feature_service in sorted(
-            self.list_feature_services(project=project),
-            key=lambda feature_service: feature_service.name,
-        ):
-            registry_dict["featureServices"].append(
-                self._message_to_sorted_dict(feature_service.to_proto())
-            )
-        for on_demand_feature_view in sorted(
-            self.list_on_demand_feature_views(project=project),
-            key=lambda on_demand_feature_view: on_demand_feature_view.name,
-        ):
-            odfv_dict = self._message_to_sorted_dict(on_demand_feature_view.to_proto())
-
-            odfv_dict["spec"]["userDefinedFunction"][
-                "body"
-            ] = on_demand_feature_view.udf_string
-            registry_dict["onDemandFeatureViews"].append(odfv_dict)
-        for request_feature_view in sorted(
-            self.list_request_feature_views(project=project),
-            key=lambda request_feature_view: request_feature_view.name,
-        ):
-            registry_dict["requestFeatureViews"].append(
-                self._message_to_sorted_dict(request_feature_view.to_proto())
-            )
-        for saved_dataset in sorted(
-            self.list_saved_datasets(project=project), key=lambda item: item.name
-        ):
-            registry_dict["savedDatasets"].append(
-                self._message_to_sorted_dict(saved_dataset.to_proto())
-            )
-        for infra_object in sorted(self.get_infra(project=project).infra_objects):
-            registry_dict["infra"].append(
-                self._message_to_sorted_dict(infra_object.to_proto())
-            )
-        return registry_dict
-
-
 def _get_project_metadata(
     registry_proto: Optional[RegistryProto], project: str
 ) -> Optional[ProjectMetadataProto]:
@@ -776,10 +164,6 @@ def _init_project_metadata(cached_registry_proto: RegistryProto, project: str):
 
 
 class Registry(BaseRegistry):
-    """
-    Registry: A registry allows for the management and persistence of feature definitions and related metadata.
-    """
-
     def apply_user_metadata(
         self,
         project: str,
@@ -864,14 +248,6 @@ class Registry(BaseRegistry):
             self._registry_store.update_registry_proto(registry_proto)
 
     def update_infra(self, infra: Infra, project: str, commit: bool = True):
-        """
-        Updates the stored Infra object.
-
-        Args:
-            infra: The new Infra object to be stored.
-            project: Feast project that the Infra object refers to
-            commit: Whether the change should be persisted immediately
-        """
         self._prepare_registry_for_changes(project)
         assert self.cached_registry_proto
 
@@ -880,30 +256,12 @@ class Registry(BaseRegistry):
             self.commit()
 
     def get_infra(self, project: str, allow_cache: bool = False) -> Infra:
-        """
-        Retrieves the stored Infra object.
-
-        Args:
-            project: Feast project that the Infra object refers to
-            allow_cache: Whether to allow returning this entity from a cached registry
-
-        Returns:
-            The stored Infra object.
-        """
         registry_proto = self._get_registry_proto(
             project=project, allow_cache=allow_cache
         )
         return Infra.from_proto(registry_proto.infra)
 
     def apply_entity(self, entity: Entity, project: str, commit: bool = True):
-        """
-        Registers a single entity with Feast
-
-        Args:
-            entity: Entity that will be registered
-            project: Feast project that this entity belongs to
-            commit: Whether the change should be persisted immediately
-        """
         entity.is_valid()
 
         now = datetime.utcnow()
@@ -931,16 +289,6 @@ class Registry(BaseRegistry):
             self.commit()
 
     def list_entities(self, project: str, allow_cache: bool = False) -> List[Entity]:
-        """
-        Retrieve a list of entities from the registry
-
-        Args:
-            allow_cache: Whether to allow returning entities from a cached registry
-            project: Filter entities based on project name
-
-        Returns:
-            List of entities
-        """
         registry_proto = self._get_registry_proto(
             project=project, allow_cache=allow_cache
         )
@@ -953,16 +301,6 @@ class Registry(BaseRegistry):
     def list_data_sources(
         self, project: str, allow_cache: bool = False
     ) -> List[DataSource]:
-        """
-        Retrieve a list of data sources from the registry
-
-        Args:
-            project: Filter data source based on project name
-            allow_cache: Whether to allow returning data sources from a cached registry
-
-        Returns:
-            List of data sources
-        """
         registry_proto = self._get_registry_proto(
             project=project, allow_cache=allow_cache
         )
@@ -975,14 +313,6 @@ class Registry(BaseRegistry):
     def apply_data_source(
         self, data_source: DataSource, project: str, commit: bool = True
     ):
-        """
-        Registers a single data source with Feast
-
-        Args:
-            data_source: A data source that will be registered
-            project: Feast project that this data source belongs to
-            commit: Whether to immediately commit to the registry
-        """
         registry = self._prepare_registry_for_changes(project)
         for idx, existing_data_source_proto in enumerate(registry.data_sources):
             if existing_data_source_proto.name == data_source.name:
@@ -1000,14 +330,6 @@ class Registry(BaseRegistry):
             self.commit()
 
     def delete_data_source(self, name: str, project: str, commit: bool = True):
-        """
-        Deletes a data source or raises an exception if not found.
-
-        Args:
-            name: Name of data source
-            project: Feast project that this data source belongs to
-            commit: Whether the change should be persisted immediately
-        """
         self._prepare_registry_for_changes(project)
         assert self.cached_registry_proto
 
@@ -1024,13 +346,6 @@ class Registry(BaseRegistry):
     def apply_feature_service(
         self, feature_service: FeatureService, project: str, commit: bool = True
     ):
-        """
-        Registers a single feature service with Feast
-
-        Args:
-            feature_service: A feature service that will be registered
-            project: Feast project that this entity belongs to
-        """
         now = datetime.utcnow()
         if not feature_service.created_timestamp:
             feature_service.created_timestamp = now
@@ -1055,17 +370,6 @@ class Registry(BaseRegistry):
     def list_feature_services(
         self, project: str, allow_cache: bool = False
     ) -> List[FeatureService]:
-        """
-        Retrieve a list of feature services from the registry
-
-        Args:
-            allow_cache: Whether to allow returning entities from a cached registry
-            project: Filter entities based on project name
-
-        Returns:
-            List of feature services
-        """
-
         registry = self._get_registry_proto(project=project, allow_cache=allow_cache)
         feature_services = []
         for feature_service_proto in registry.feature_services:
@@ -1078,18 +382,6 @@ class Registry(BaseRegistry):
     def get_feature_service(
         self, name: str, project: str, allow_cache: bool = False
     ) -> FeatureService:
-        """
-        Retrieves a feature service.
-
-        Args:
-            name: Name of feature service
-            project: Feast project that this feature service belongs to
-            allow_cache: Whether to allow returning this feature service from a cached registry
-
-        Returns:
-            Returns either the specified feature service, or raises an exception if
-            none is found
-        """
         registry = self._get_registry_proto(project=project, allow_cache=allow_cache)
 
         for feature_service_proto in registry.feature_services:
@@ -1101,18 +393,6 @@ class Registry(BaseRegistry):
         raise FeatureServiceNotFoundException(name, project=project)
 
     def get_entity(self, name: str, project: str, allow_cache: bool = False) -> Entity:
-        """
-        Retrieves an entity.
-
-        Args:
-            name: Name of entity
-            project: Feast project that this entity belongs to
-            allow_cache: Whether to allow returning this entity from a cached registry
-
-        Returns:
-            Returns either the specified entity, or raises an exception if
-            none is found
-        """
         registry_proto = self._get_registry_proto(
             project=project, allow_cache=allow_cache
         )
@@ -1124,14 +404,6 @@ class Registry(BaseRegistry):
     def apply_feature_view(
         self, feature_view: BaseFeatureView, project: str, commit: bool = True
     ):
-        """
-        Registers a single feature view with Feast
-
-        Args:
-            feature_view: Feature view that will be registered
-            project: Feast project that this feature view belongs to
-            commit: Whether the change should be persisted immediately
-        """
         feature_view.ensure_valid()
 
         now = datetime.utcnow()
@@ -1188,16 +460,6 @@ class Registry(BaseRegistry):
     def list_stream_feature_views(
         self, project: str, allow_cache: bool = False
     ) -> List[StreamFeatureView]:
-        """
-        Retrieve a list of stream feature views from the registry
-
-        Args:
-            project: Filter stream feature views based on project name
-            allow_cache: Whether to allow returning stream feature views from a cached registry
-
-        Returns:
-            List of stream feature views
-        """
         registry = self._get_registry_proto(project=project, allow_cache=allow_cache)
         stream_feature_views = []
         for stream_feature_view in registry.stream_feature_views:
@@ -1210,17 +472,6 @@ class Registry(BaseRegistry):
     def list_on_demand_feature_views(
         self, project: str, allow_cache: bool = False
     ) -> List[OnDemandFeatureView]:
-        """
-        Retrieve a list of on demand feature views from the registry
-
-        Args:
-            project: Filter on demand feature views based on project name
-            allow_cache: Whether to allow returning on demand feature views from a cached registry
-
-        Returns:
-            List of on demand feature views
-        """
-
         registry = self._get_registry_proto(project=project, allow_cache=allow_cache)
         on_demand_feature_views = []
         for on_demand_feature_view in registry.on_demand_feature_views:
@@ -1233,18 +484,6 @@ class Registry(BaseRegistry):
     def get_on_demand_feature_view(
         self, name: str, project: str, allow_cache: bool = False
     ) -> OnDemandFeatureView:
-        """
-        Retrieves an on demand feature view.
-
-        Args:
-            name: Name of on demand feature view
-            project: Feast project that this on demand feature view belongs to
-            allow_cache: Whether to allow returning this on demand feature view from a cached registry
-
-        Returns:
-            Returns either the specified on demand feature view, or raises an exception if
-            none is found
-        """
         registry = self._get_registry_proto(project=project, allow_cache=allow_cache)
 
         for on_demand_feature_view in registry.on_demand_feature_views:
@@ -1258,17 +497,6 @@ class Registry(BaseRegistry):
     def get_data_source(
         self, name: str, project: str, allow_cache: bool = False
     ) -> DataSource:
-        """
-        Retrieves a data source.
-
-        Args:
-            name: Name of data source
-            project: Feast project that this data source belongs to
-            allow_cache: Whether to allow returning this data source from a cached registry
-
-        Returns:
-            Returns either the specified data source, or raises an exception if none is found
-        """
         registry = self._get_registry_proto(project=project, allow_cache=allow_cache)
 
         for data_source in registry.data_sources:
@@ -1284,16 +512,6 @@ class Registry(BaseRegistry):
         end_date: datetime,
         commit: bool = True,
     ):
-        """
-        Updates materialization intervals tracked for a single feature view in Feast
-
-        Args:
-            feature_view: Feature view that will be updated with an additional materialization interval tracked
-            project: Feast project that this feature view belongs to
-            start_date (datetime): Start date of the materialization interval to track
-            end_date (datetime): End date of the materialization interval to track
-            commit: Whether the change should be persisted immediately
-        """
         self._prepare_registry_for_changes(project)
         assert self.cached_registry_proto
 
@@ -1348,16 +566,6 @@ class Registry(BaseRegistry):
     def list_feature_views(
         self, project: str, allow_cache: bool = False
     ) -> List[FeatureView]:
-        """
-        Retrieve a list of feature views from the registry
-
-        Args:
-            allow_cache: Allow returning feature views from the cached registry
-            project: Filter feature views based on project name
-
-        Returns:
-            List of feature views
-        """
         registry_proto = self._get_registry_proto(
             project=project, allow_cache=allow_cache
         )
@@ -1368,17 +576,6 @@ class Registry(BaseRegistry):
         return feature_views
 
     def get_request_feature_view(self, name: str, project: str):
-        """
-        Retrieves a feature view.
-
-        Args:
-            name: Name of feature view
-            project: Feast project that this feature view belongs to
-
-        Returns:
-            Returns either the specified feature view, or raises an exception if
-            none is found
-        """
         registry_proto = self._get_registry_proto(project=project, allow_cache=False)
         for feature_view_proto in registry_proto.feature_views:
             if (
@@ -1391,16 +588,6 @@ class Registry(BaseRegistry):
     def list_request_feature_views(
         self, project: str, allow_cache: bool = False
     ) -> List[RequestFeatureView]:
-        """
-        Retrieve a list of request feature views from the registry
-
-        Args:
-            allow_cache: Allow returning feature views from the cached registry
-            project: Filter feature views based on project name
-
-        Returns:
-            List of feature views
-        """
         registry_proto = self._get_registry_proto(
             project=project, allow_cache=allow_cache
         )
@@ -1415,18 +602,6 @@ class Registry(BaseRegistry):
     def get_feature_view(
         self, name: str, project: str, allow_cache: bool = False
     ) -> FeatureView:
-        """
-        Retrieves a feature view.
-
-        Args:
-            name: Name of feature view
-            project: Feast project that this feature view belongs to
-            allow_cache: Allow returning feature view from the cached registry
-
-        Returns:
-            Returns either the specified feature view, or raises an exception if
-            none is found
-        """
         registry_proto = self._get_registry_proto(
             project=project, allow_cache=allow_cache
         )
@@ -1441,18 +616,6 @@ class Registry(BaseRegistry):
     def get_stream_feature_view(
         self, name: str, project: str, allow_cache: bool = False
     ) -> StreamFeatureView:
-        """
-        Retrieves a stream feature view.
-
-        Args:
-            name: Name of stream feature view
-            project: Feast project that this stream feature view belongs to
-            allow_cache: Allow returning feature view from the cached registry
-
-        Returns:
-            Returns either the specified feature view, or raises an exception if
-            none is found
-        """
         registry_proto = self._get_registry_proto(
             project=project, allow_cache=allow_cache
         )
@@ -1465,14 +628,6 @@ class Registry(BaseRegistry):
         raise FeatureViewNotFoundException(name, project)
 
     def delete_feature_service(self, name: str, project: str, commit: bool = True):
-        """
-        Deletes a feature service or raises an exception if not found.
-
-        Args:
-            name: Name of feature service
-            project: Feast project that this feature service belongs to
-            commit: Whether the change should be persisted immediately
-        """
         self._prepare_registry_for_changes(project)
         assert self.cached_registry_proto
 
@@ -1490,14 +645,6 @@ class Registry(BaseRegistry):
         raise FeatureServiceNotFoundException(name, project)
 
     def delete_feature_view(self, name: str, project: str, commit: bool = True):
-        """
-        Deletes a feature view or raises an exception if not found.
-
-        Args:
-            name: Name of feature view
-            project: Feast project that this feature view belongs to
-            commit: Whether the change should be persisted immediately
-        """
         self._prepare_registry_for_changes(project)
         assert self.cached_registry_proto
 
@@ -1552,14 +699,6 @@ class Registry(BaseRegistry):
         raise FeatureViewNotFoundException(name, project)
 
     def delete_entity(self, name: str, project: str, commit: bool = True):
-        """
-        Deletes an entity or raises an exception if not found.
-
-        Args:
-            name: Name of entity
-            project: Feast project that this entity belongs to
-            commit: Whether the change should be persisted immediately
-        """
         self._prepare_registry_for_changes(project)
         assert self.cached_registry_proto
 
@@ -1583,14 +722,6 @@ class Registry(BaseRegistry):
         project: str,
         commit: bool = True,
     ):
-        """
-        Stores a saved dataset metadata with Feast
-
-        Args:
-            saved_dataset: SavedDataset that will be added / updated to registry
-            project: Feast project that this dataset belongs to
-            commit: Whether the change should be persisted immediately
-        """
         now = datetime.utcnow()
         if not saved_dataset.created_timestamp:
             saved_dataset.created_timestamp = now
@@ -1618,18 +749,6 @@ class Registry(BaseRegistry):
     def get_saved_dataset(
         self, name: str, project: str, allow_cache: bool = False
     ) -> SavedDataset:
-        """
-        Retrieves a saved dataset.
-
-        Args:
-            name: Name of dataset
-            project: Feast project that this dataset belongs to
-            allow_cache: Whether to allow returning this dataset from a cached registry
-
-        Returns:
-            Returns either the specified SavedDataset, or raises an exception if
-            none is found
-        """
         registry_proto = self._get_registry_proto(
             project=project, allow_cache=allow_cache
         )
@@ -1644,16 +763,6 @@ class Registry(BaseRegistry):
     def list_saved_datasets(
         self, project: str, allow_cache: bool = False
     ) -> List[SavedDataset]:
-        """
-        Retrieves a list of all saved datasets in specified project
-
-        Args:
-            project: Feast project
-            allow_cache: Whether to allow returning this dataset from a cached registry
-
-        Returns:
-            Returns the list of SavedDatasets
-        """
         registry_proto = self._get_registry_proto(
             project=project, allow_cache=allow_cache
         )
@@ -1669,14 +778,6 @@ class Registry(BaseRegistry):
         project: str,
         commit: bool = True,
     ):
-        """
-        Persist a validation reference
-
-        Args:
-            validation_reference: ValidationReference that will be added / updated to registry
-            project: Feast project that this dataset belongs to
-            commit: Whether the change should be persisted immediately
-        """
         validation_reference_proto = validation_reference.to_proto()
         validation_reference_proto.project = project
 
@@ -1698,18 +799,6 @@ class Registry(BaseRegistry):
     def get_validation_reference(
         self, name: str, project: str, allow_cache: bool = False
     ) -> ValidationReference:
-        """
-        Retrieves a validation reference.
-
-        Args:
-            name: Name of dataset
-            project: Feast project that this dataset belongs to
-            allow_cache: Whether to allow returning this dataset from a cached registry
-
-        Returns:
-            Returns either the specified ValidationReference, or raises an exception if
-            none is found
-        """
         registry_proto = self._get_registry_proto(
             project=project, allow_cache=allow_cache
         )
@@ -1722,14 +811,6 @@ class Registry(BaseRegistry):
         raise ValidationReferenceNotFound(name, project=project)
 
     def delete_validation_reference(self, name: str, project: str, commit: bool = True):
-        """
-        Deletes a validation reference or raises an exception if not found.
-
-        Args:
-            name: Name of validation reference
-            project: Feast project that this object belongs to
-            commit: Whether the change should be persisted immediately
-        """
         registry_proto = self._prepare_registry_for_changes(project)
         for idx, existing_validation_reference in enumerate(
             registry_proto.validation_references
