+++ sdk/python/tests/foo_provider.py
@@ -9,7 +9,7 @@ from tqdm import tqdm
 from feast import Entity, FeatureService, FeatureView, RepoConfig
 from feast.infra.offline_stores.offline_store import RetrievalJob
 from feast.infra.provider import Provider
-from feast.infra.registry.registry import BaseRegistry
+from feast.infra.registry.base_registry import BaseRegistry
 from feast.protos.feast.types.EntityKey_pb2 import EntityKey as EntityKeyProto
 from feast.protos.feast.types.Value_pb2 import Value as ValueProto
 from feast.saved_dataset import SavedDataset
